# Conflicts:
#	src/Adonan/BahanRotiTawarPerPcs.java

+++ src/Adonan/TotalBeratBahan.java
@@ -1,5 +1,5 @@
 package Adonan;
 
 public interface TotalBeratBahan {
-    public int getTotalBerat();
+    public double getTotalBerat();
 }
